fingerprinting/requirements.txt
fpdf==1.7.2
Pillow==8.4.0
pytest==6.2.5
click==8.0.3
prompt==0.4.1